% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFormat.R
\name{dataFormat}
\alias{dataFormat}
\title{Format data for plots}
\usage{
dataFormat(x, measurevar = "residuals", categories = c(subject = "sub", time
  = "isi", type = "type", group = "music"))
}
\arguments{
\item{x}{data frame to format}

\item{measurevar}{dependent variable we are calculating values for}

\item{categories}{list of grouping variables}

\item{subject}{name of subject id}

\item{time}{grouping variable for time (ie. isi)}

\item{type}{grouping variable for stimulus type}

\item{group}{grouping variable for experimental group}
}
\value{
correctly formatted data frame with means for each subject in each category
}
\description{
Format data for plots
}
\examples{
   dataFormat(x,measurevar="residuals",categories=c(subject="sub",time="isi",type="type",group="music"))
}
