% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorBars.R
\name{errorBars}
\alias{errorBars}
\title{Error bar calculations}
\usage{
errorBars(mydata = "mydata", measurevar = "measurevar", groupvars = c(time
  = "time", type = "type", group = "group"), na.rm = FALSE,
  conf.interval = 0.95, .drop = TRUE)
}
\arguments{
\item{mydata}{the data frame we will calculate error bars for}

\item{measurevar}{the dependent variable we are calculating error bars for}

\item{groupvars}{grouping variables for error bars}

\item{time}{time variable (ie. isi, interval)}

\item{type}{type variable (ie. stimulus category)}

\item{group}{group variable (ie. condition)}
}
\value{
N, mean, standard deviation, standard error, 95% confidence interval
}
\description{
Error bar calculations
}
\examples{
   errorBars(mydata="mydata", measurevar="measurevar", groupvars=c(time="time",type="type",group="group"))
}
