% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePlot.R
\name{makePlot}
\alias{makePlot}
\title{Make super cool plots}
\usage{
makePlot(plotData = "plotData", time = "time", measurevar = "measurevar",
  type = "type", group = "group", xaxis = "interstimulus interval (sec)",
  yaxis = "residual reaction time (ms)", se = "se", barWidth = 0.03,
  legtitle = "Word Type", legcolors = c("firebrick1", "limegreen",
  "royalblue1"), leglevels = c("n", "w", "a"), leglabels = c("non-word",
  "non-associate", "associate"), plotorder = c("nomusic", "slowmusic",
  "fastmusic"), plottitles = c("No Music", "Slow Music", "Fast Music"))
}
\arguments{
\item{plotData}{formatted data frame to use for plot}

\item{time}{time variable for x-axis}

\item{measurevar}{dependent variable for y-axis}

\item{type}{grouping variable for stimulus type (what lines are split by)}

\item{group}{grouping variable for experimental group (each on separate plot)}

\item{xaxis}{label for x-axis}

\item{yaxis}{label for y-axis}

\item{se}{standard error value for error bars}

\item{barWidth}{width of error bars (default .025)}
}
\value{
awesome plot of our formatted data; includes separate lines for stimulus type and separate plots for experimental group
}
\description{
Make super cool plots
}
\examples{
   makePlot(mydata="mydata",time="time",measurevar="measurevar",type="type",group="group",xaxis="interstimulus interval (sec)",yaxis="residual reaction time (ms)",se="sem")
}
